corsproxy/requirements.txt
flask
gunicorn
requests
PyJWT
bcrypt
pytest
pytest-mock
